feat:  add conditional to verify is entity has fk on delete

+++ src/data/services/Jurisdiction/DeleteJurisdiction.service.js
@@ -1,8 +1,12 @@
-import { JurisdictionRepository } from '../../database/repositories';
+import {
+  JurisdictionRepository,
+  AgencyRepository,
+} from '../../database/repositories';
 
 export class DeleteJurisdictionService {
   async execute({ id }) {
     const repository = new JurisdictionRepository();
+    const agencyRepository = new AgencyRepository();
 
     const verifyJurisdictionExists = await repository.findJurisdictionById({
       id,
@@ -12,6 +16,17 @@ export class DeleteJurisdictionService {
       return {
         error: `Não há nenhuma esfera registrada com este ID -> ${id}.`,
       };
+
+    const verifyFk = await agencyRepository.verifyJurisdiction({
+      jurisdictionId: id,
+    });
+
+    if (verifyFk.length > 0) {
+      return {
+        error:
+          'Não foi possível excluir a Esfera pois existem Órgãos associados.',
+      };
+    }
 
     await repository.deleteJurisdiction({
       id,
