+++ src/data/services/Region/DeleteRegion.service.js
@@ -1,8 +1,9 @@
-import { RegionRepository } from '../../database/repositories';
+import { RegionRepository, CityRepository } from '../../database/repositories';
 
 export class DeleteRegionService {
   async execute({ id }) {
     const repository = new RegionRepository();
+    const cityRepository = new CityRepository();
 
     const verifyRegionExists = await repository.findRegionById({
       id,
@@ -12,6 +13,15 @@ export class DeleteRegionService {
       return {
         error: `Não existe uma região registrada com este ID -> ${id}.`,
       };
+
+    const verifyFk = await cityRepository.verifyRegion({ regionId: id });
+
+    if (verifyFk.length > 0) {
+      return {
+        error:
+          'Não foi possível excluir a Região pois existem Municípios associados.',
+      };
+    }
 
     await repository.deleteRegion({
       id,
